scripts/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
discord==2.3.2
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
instagrapi==2.1.1
multidict==6.0.5
oauthlib==3.2.2
pillow==10.2.0
praw==7.7.1
prawcore==2.4.0
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.4.0
tweepy==4.14.0
typing_extensions==4.10.0
update-checker==0.18.0
urllib3==2.2.1
websocket-client==1.7.0
yarl==1.9.4
